Applied hexagonal architecture to insert_ballots
