+++ packages/windmill/src/services/ceremonies/insert_ballots/domain.rs
@@ -3,10 +3,19 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
 use anyhow::{anyhow, Result};
+use b3::messages::artifact::Configuration;
+use b3::messages::newtypes::{TrusteeSet, MAX_TRUSTEES, NULL_TRUSTEE};
 use csv::ReaderBuilder;
 use std::{cmp::Ordering, fs::File};
-use tracing::{info, instrument};
+use strand::context::Ctx;
+use strand::signature::StrandSignaturePk;
+use tracing::{event, info, instrument, Level};
 
+/// Performs a sorted merge join on two CSV files and returns the matched ballot contents
+/// together with voter and ballot statistics.
+///
+/// Both files must be sorted lexicographically on their join key column. The function
+/// yields `(matched_ballot_contents, eligible_voters, ballots_without_voter, casted_ballots)`.
 #[instrument(skip_all, err)]
 pub fn merge_join_csv(
     ballots_file: &File,
@@ -18,13 +27,11 @@ pub fn merge_join_csv(
 ) -> Result<(Vec<String>, u64, u64, u64)> {
     info!("START merge_join_csv");
 
-    // Initialize the result vector and counters
     let mut result = Vec::new();
     let mut ballots_without_voter: u64 = 0;
     let mut elegible_voters: u64 = 0;
     let mut casted_ballots: u64 = 0;
 
-    // Assume the CSV files do not have headers.
     let mut ballots_reader = ReaderBuilder::new()
         .has_headers(false)
         .from_reader(ballots_file);
@@ -32,17 +39,13 @@ pub fn merge_join_csv(
         .has_headers(false)
         .from_reader(voters_file);
 
-    // Create iterators over CSV records.
     let mut ballots_iterator = ballots_reader.records();
     let mut voters_iterator = voters_reader.records();
 
-    // Read the first record from each file.
     let mut ballots_record = ballots_iterator.next();
     let mut voters_record = voters_iterator.next();
 
-    // Continue while both files still have records.
     while ballots_record.is_some() && voters_record.is_some() {
-        // Unwrap the current records.
         let Some(Ok(ballot)) = ballots_record.as_ref() else {
             ballots_record = ballots_iterator.next();
             continue;
@@ -52,72 +55,53 @@ pub fn merge_join_csv(
             continue;
         };
 
-        // Extract the ballot join key.
         let Some(ballot_voter_id) = ballot.get(ballots_voter_id_index) else {
-            // Advance ballots file.
             ballots_record = ballots_iterator.next();
             continue;
         };
-        // Ignore ballots with an empty key.
         if ballot_voter_id.is_empty() {
             ballots_record = ballots_iterator.next();
             continue;
         }
 
-        // Extract the voter join key.
         let Some(voter_id) = voter.get(voters_id_index) else {
-            // Advance voters file.
             voters_record = voters_iterator.next();
             continue;
         };
-        // Ignore users with an empty key.
         if voter_id.is_empty() {
             voters_record = voters_iterator.next();
             continue;
         }
 
-        // --- Delegate Count Logic ---
-        // This block runs only if the delegate feature is enabled.
-        // If parsing fails, we skip the voter record and continue the loop.
         let delegate_count: usize = if let Some(index) = delegate_count_index {
             let Some(delegate_count_str) = voter.get(index) else {
-                // Failed to get field, advance voter and continue loop
                 voters_record = voters_iterator.next();
                 continue;
             };
             if delegate_count_str.is_empty() {
-                // Empty field, advance voter and continue loop
                 voters_record = voters_iterator.next();
                 continue;
             }
             let Ok(count) = delegate_count_str.parse() else {
-                // Invalid number, advance voter and continue loop
                 voters_record = voters_iterator.next();
                 continue;
             };
             count
         } else {
-            // Delegate feature is disabled, default to 0.
             0
         };
-        // --- End Delegate Count Logic ---
 
-        // Compare the join keys lexicographically.
         match ballot_voter_id.cmp(voter_id) {
             Ordering::Less => {
-                // If the ballot has no voter.
                 ballots_without_voter += 1;
-                // Advance ballots file.
                 ballots_record = ballots_iterator.next();
                 casted_ballots += 1;
             }
             Ordering::Greater => {
-                // Advance voters file.
                 voters_record = voters_iterator.next();
                 elegible_voters += 1;
             }
             Ordering::Equal => {
-                // Match found.
                 let ballot_content = ballot.get(ballots_content_index).ok_or_else(|| {
                     anyhow!(
                         "Output column index {} out of bounds in file1",
@@ -125,30 +109,23 @@ pub fn merge_join_csv(
                     )
                 })?;
 
-                // Add the voter's own ballot.
                 result.push(ballot_content.to_string());
-
-                // Add delegates if any (if delegate_count was 0, this does nothing).
                 result.extend(std::iter::repeat(ballot_content.to_string()).take(delegate_count));
 
-                // Advance both iterators.
                 ballots_record = ballots_iterator.next();
                 voters_record = voters_iterator.next();
 
-                // Count the voter's ballot (1) + all their delegated ballots.
                 casted_ballots += 1 + (delegate_count as u64);
                 elegible_voters += 1;
             }
         }
     }
 
-    // Count the rest of the voters
     while voters_record.is_some() {
         elegible_voters += 1;
         voters_record = voters_iterator.next();
     }
 
-    // Count the rest of the ballots
     while ballots_record.is_some() {
         casted_ballots += 1;
         ballots_without_voter += 1;
@@ -165,13 +142,93 @@ pub fn merge_join_csv(
     ))
 }
 
+/// Maps trustee public keys to their 1-indexed positions in the board configuration's trustee
+/// list and returns a fixed-size `TrusteeSet` array.
+#[instrument(skip_all)]
+pub fn generate_trustee_set<C: Ctx>(
+    configuration: &Configuration<C>,
+    trustee_pks: Vec<StrandSignaturePk>,
+) -> TrusteeSet {
+    let result = find_trustee_positions(&configuration.trustees, trustee_pks);
+    event!(Level::INFO, "TrusteeSet: {:?}", result);
+    result
+}
+
+/// Returns a `TrusteeSet` array where each slot holds the 1-based position of the corresponding
+/// `trustee_pks` entry within `trustees`, or `NULL_TRUSTEE` if not found.
+fn find_trustee_positions(
+    trustees: &[StrandSignaturePk],
+    trustee_pks: Vec<StrandSignaturePk>,
+) -> TrusteeSet {
+    let mut selected_trustees: TrusteeSet = [NULL_TRUSTEE; MAX_TRUSTEES];
+    let trustee_ids: Vec<usize> = trustee_pks
+        .into_iter()
+        .map(|trustee_pk| {
+            let position = trustees.iter().position(|trustee| trustee == &trustee_pk);
+            match position {
+                Some(value) => value + 1,
+                None => NULL_TRUSTEE,
+            }
+        })
+        .collect();
+    for i in 0..trustee_ids.len() {
+        selected_trustees[i] = trustee_ids[i];
+    }
+    selected_trustees
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+    use b3::messages::newtypes::NULL_TRUSTEE;
     use std::io::Write;
+    use strand::signature::{StrandSignaturePk, StrandSignatureSk};
     use tempfile::NamedTempFile;
 
-    /// Helper function to run tests for `merge_join_csv` (non-delegate mode).
+    fn make_pk() -> StrandSignaturePk {
+        let sk = StrandSignatureSk::gen().unwrap();
+        StrandSignaturePk::from_sk(&sk).unwrap()
+    }
+
+    #[test]
+    fn test_all_trustees_found() {
+        let pks: Vec<StrandSignaturePk> = (0..3).map(|_| make_pk()).collect();
+        let result = find_trustee_positions(&pks, pks.clone());
+        assert_eq!(result[0], 1);
+        assert_eq!(result[1], 2);
+        assert_eq!(result[2], 3);
+        assert!(result[3..].iter().all(|&x| x == NULL_TRUSTEE));
+    }
+
+    #[test]
+    fn test_trustee_not_in_config() {
+        let config_pks: Vec<StrandSignaturePk> = (0..2).map(|_| make_pk()).collect();
+        let unknown_pk = make_pk();
+        let result = find_trustee_positions(&config_pks, vec![config_pks[0].clone(), unknown_pk]);
+        assert_eq!(result[0], 1);
+        assert_eq!(result[1], NULL_TRUSTEE);
+        assert!(result[2..].iter().all(|&x| x == NULL_TRUSTEE));
+    }
+
+    #[test]
+    fn test_empty_pks() {
+        let config_pks: Vec<StrandSignaturePk> = (0..2).map(|_| make_pk()).collect();
+        let result = find_trustee_positions(&config_pks, vec![]);
+        assert!(result.iter().all(|&x| x == NULL_TRUSTEE));
+    }
+
+    #[test]
+    fn test_out_of_order_pks() {
+        let pks: Vec<StrandSignaturePk> = (0..3).map(|_| make_pk()).collect();
+        // Provide pks in reverse order
+        let reversed = vec![pks[2].clone(), pks[0].clone(), pks[1].clone()];
+        let result = find_trustee_positions(&pks, reversed);
+        assert_eq!(result[0], 3);
+        assert_eq!(result[1], 1);
+        assert_eq!(result[2], 2);
+        assert!(result[3..].iter().all(|&x| x == NULL_TRUSTEE));
+    }
+
     fn run_merge_join_test(
         ballots_csv: &str,
         users_csv: &str,
@@ -187,17 +244,14 @@ mod tests {
         let ballots_ro = ballots_file.reopen()?;
         let users_ro = users_file.reopen()?;
 
-        // Assumes standard test indexes:
-        // ballots_voter_id_index=0, voters_id_index=0, ballots_content_index=1
-        // Pass `None` for delegate_count_index to run in standard mode.
         let (ballot_contents, elegible_voters, ballots_without_voter, casted_ballots) =
             merge_join_csv(
                 &ballots_ro,
                 &users_ro,
-                0,    // ballots_voter_id_index
-                0,    // voters_id_index
-                1,    // ballots_content_index
-                None, // delegate_count_index
+                0,
+                0,
+                1,
+                None,
             )?;
         Ok((
             ballot_contents,
@@ -209,7 +263,6 @@ mod tests {
 
     #[test]
     fn test_basic_auditable_ballot() -> Result<()> {
-        // user_C's ballot should be counted as auditable as they are not in the users file.
         let ballots = "user_A,content_A\nuser_B,content_B\nuser_C,content_C";
         let users = "user_A\nuser_B";
         let (_, elegible_voters, ballots_without_voter, casted_ballots) =
@@ -222,7 +275,6 @@ mod tests {
 
     #[test]
     fn test_no_auditable_ballots_all_match() -> Result<()> {
-        // All users who voted are in the enabled users list.
         let ballots = "user_A,content_A\nuser_B,content_B";
         let users = "user_A\nuser_B";
         let (_, elegible_voters, ballots_without_voter, casted_ballots) =
@@ -235,8 +287,6 @@ mod tests {
 
     #[test]
     fn test_auditable_ballots_at_end_of_file() -> Result<()> {
-        // This specifically tests the bug fix. user_C and user_D's ballots are after
-        // the last user in the users file. The old buggy code would miss these.
         let ballots = "user_A,content_A\nuser_C,content_C\nuser_D,content_D";
         let users = "user_A\nuser_B";
         let (_, elegible_voters, ballots_without_voter, casted_ballots) =
@@ -249,7 +299,6 @@ mod tests {
 
     #[test]
     fn test_empty_ballot_file() -> Result<()> {
-        // If there are no ballots, the count must be 0.
         let ballots = "";
         let users = "user_A\nuser_B";
         let (_, elegible_voters, ballots_without_voter, casted_ballots) =
@@ -262,7 +311,6 @@ mod tests {
 
     #[test]
     fn test_empty_enabled_users_file() -> Result<()> {
-        // If the enabled users list is empty, all ballots should be counted as auditable.
         let ballots = "user_A,content_A\nuser_B,content_B\nuser_C,content_C";
         let users = "";
         let (_, elegible_voters, ballots_without_voter, casted_ballots) =
@@ -275,7 +323,6 @@ mod tests {
 
     #[test]
     fn test_both_files_empty() -> Result<()> {
-        // If both files are empty, the count is 0.
         let ballots = "";
         let users = "";
         let (_, elegible_voters, ballots_without_voter, casted_ballots) =
@@ -288,12 +335,6 @@ mod tests {
 
     #[test]
     fn test_mixed_scenario_with_gaps() -> Result<()> {
-        // A more complex real-world scenario.
-        // user_A: match
-        // user_C: auditable
-        // user_E: match
-        // user_F: auditable
-        // user_H: auditable
         let ballots = "user_A,content_A\nuser_C,content_C\nuser_E,content_E\nuser_F,content_F\nuser_H,content_H";
         let users = "user_A\nuser_B\nuser_D\nuser_E\nuser_G";
         let (_, elegible_voters, ballots_without_voter, casted_ballots) =
@@ -306,13 +347,6 @@ mod tests {
 
     #[test]
     fn test_handles_malformed_but_consistent_columns() -> Result<()> {
-        // This test has consistent column counts, but contains invalid data
-        // like empty strings for keys, which should be skipped by the function's logic.
-        //
-        // - Row 1: ``,content_A` -> Skipped (empty key1)
-        // - Row 2: `user_B,content_B` -> VALID AUDITABLE BALLOT
-        // - Row 3: `user_C,content_C` -> VALID AUDITABLE BALLOT
-        // - Row 4: `user_D,content_D` -> Valid, but matches user_D, so not auditable.
         let ballots = ",content_A\nuser_B,content_B\nuser_C,content_C\nuser_D,content_D";
         let users = "user_A\nuser_D";
         let (_, elegible_voters, ballots_without_voter, casted_ballots) =
@@ -326,30 +360,22 @@ mod tests {
     #[test]
     fn test_large_scale_auditable_count() -> Result<()> {
         const TOTAL_ENTRIES: u64 = 500;
-        const EXPECTED_AUDITABLE_COUNT: u64 = (TOTAL_ENTRIES / 2) as u64; // We will add only even users, so odds are auditable.
+        const EXPECTED_AUDITABLE_COUNT: u64 = (TOTAL_ENTRIES / 2) as u64;
 
         let mut ballots_csv = String::new();
         let mut users_csv = String::new();
 
-        // Generate hundreds of "random-like" but deterministic entries.
-        // The user IDs are padded with zeros to ensure correct lexicographical sorting.
         for i in 0..TOTAL_ENTRIES {
             let user_id = format!("user-{:04}", i);
-
-            // 1. Add a ballot for every single user.
             ballots_csv.push_str(&format!("{},content_{}\n", user_id, i));
-
-            // 2. Add only users with an even index to the "enabled users" file.
             if i % 2 == 0 {
                 users_csv.push_str(&format!("{}\n", user_id));
             }
         }
 
-        // Run the test with the generated data.
         let (_, elegible_voters, ballots_without_voter, casted_ballots) =
             run_merge_join_test(&ballots_csv, &users_csv)?;
 
-        // 3. The function should count exactly half the entries—the ones we omitted (the odds).
         assert_eq!(elegible_voters, EXPECTED_AUDITABLE_COUNT);
         assert_eq!(ballots_without_voter, EXPECTED_AUDITABLE_COUNT);
         assert_eq!(casted_ballots, TOTAL_ENTRIES);
@@ -359,7 +385,6 @@ mod tests {
 
     #[test]
     fn test_merge_join_basic_join() -> Result<()> {
-        // Both ballots have a corresponding enabled user, so both contents should be returned.
         let ballots = "user_A,content_A\nuser_B,content_B";
         let users = "user_A\nuser_B";
         let (result, _, _, _) = run_merge_join_test(ballots, users)?;
@@ -369,7 +394,6 @@ mod tests {
 
     #[test]
     fn test_merge_join_partial_join() -> Result<()> {
-        // Only user_A exists in both files. user_C's ballot should be ignored.
         let ballots = "user_A,content_A\nuser_C,content_C";
         let users = "user_A\nuser_B";
         let (result, _, _, _) = run_merge_join_test(ballots, users)?;
@@ -379,7 +403,6 @@ mod tests {
 
     #[test]
     fn test_merge_join_no_matches() -> Result<()> {
-        // No common users between the two files.
         let ballots = "user_A,content_A";
         let users = "user_B\nuser_C";
         let (result, _, _, _) = run_merge_join_test(ballots, users)?;
@@ -389,11 +412,8 @@ mod tests {
 
     #[test]
     fn test_merge_join_ignores_empty_keys() -> Result<()> {
-        // *** CRITICAL TEST ***
-        // This confirms the fix for the empty key bug.
-        // The empty keys in both files should NOT result in a successful join.
         let ballots = "user_A,content_A\n,bad_content";
-        let users = "user_A\n"; // Note the empty user record
+        let users = "user_A\n";
         let (result, _, _, _) = run_merge_join_test(ballots, users)?;
         assert_eq!(result, vec!["content_A"]);
         Ok(())
@@ -401,8 +421,6 @@ mod tests {
 
     #[test]
     fn test_merge_join_handles_malformed_csv() -> Result<()> {
-        // This confirms the function skips malformed rows gracefully.
-        // The "user_B" record is missing columns and should be ignored.
         let ballots = "user_A,content_A\nuser_B\nuser_C,content_C";
         let users = "user_A\nuser_C";
         let (result, _, _, _) = run_merge_join_test(ballots, users)?;
@@ -412,7 +430,6 @@ mod tests {
 
     #[test]
     fn test_merge_join_large_scale() -> Result<()> {
-        // Stress test with a larger data set.
         const TOTAL_ENTRIES: i32 = 500;
         const EXPECTED_JOIN_COUNT: usize = (TOTAL_ENTRIES / 2) as usize;
 
@@ -421,9 +438,7 @@ mod tests {
 
         for i in 0..TOTAL_ENTRIES {
             let user_id = format!("user-{:04}", i);
-            // Add a ballot for every user.
             ballots_csv.push_str(&format!("{},content_{}\n", user_id, i));
-            // Add only even-indexed users to the enabled list.
             if i % 2 == 0 {
                 users_csv.push_str(&format!("{}\n", user_id));
             }
@@ -431,29 +446,17 @@ mod tests {
 
         let (result, _, _, _) = run_merge_join_test(&ballots_csv, &users_csv)?;
 
-        // The function should join and return only the 250 ballots from the even users.
         assert_eq!(result.len(), EXPECTED_JOIN_COUNT);
-        // Spot check the first and last expected content.
         assert_eq!(result.first().unwrap(), "content_0");
         assert_eq!(result.last().unwrap(), "content_498");
 
         Ok(())
     }
 
-    /// Helper that writes the two CSV strings to temporary files,
-    /// reopens them for reading and then calls `merge_join_csv` in delegate mode.
-    ///
-    /// The index arguments are the *standard* test indexes used by the
-    /// original function:
-    ///   ballots_voter_id_index = 0
-    ///   voters_id_index        = 0
-    ///   ballots_content_index  = 1
-    ///   delegate_count_index   = 1
     fn run_merge_join_delegates_test(
         ballots_csv: &str,
         voters_csv: &str,
     ) -> Result<(Vec<String>, u64, u64, u64)> {
-        // Write the CSV strings to temporary files
         let mut ballots_file = NamedTempFile::new()?;
         write!(ballots_file, "{}", ballots_csv)?;
         ballots_file.flush()?;
@@ -462,20 +465,17 @@ mod tests {
         write!(voters_file, "{}", voters_csv)?;
         voters_file.flush()?;
 
-        // Reopen the files read‑only – the original function expects `&File`
         let ballots_ro = ballots_file.reopen()?;
         let voters_ro = voters_file.reopen()?;
 
-        // Call the function under test
-        // Pass `Some(1)` for delegate_count_index to run in delegate mode.
         let (ballot_contents, elegible_voters, ballots_without_voter, casted_ballots) =
             merge_join_csv(
                 &ballots_ro,
                 &voters_ro,
-                /* ballots_voter_id_index   */ 0,
-                /* voters_id_index        */ 0,
-                /* ballots_content_index  */ 1,
-                /* delegate_count_index   */ Some(1),
+                0,
+                0,
+                1,
+                Some(1),
             )?;
 
         Ok((
@@ -486,12 +486,8 @@ mod tests {
         ))
     }
 
-    /// ------------------------------------------------------------------
-    /// 1. Basic delegate counting
-    /// ------------------------------------------------------------------
     #[test]
     fn test_basic_delegate_counts() -> Result<()> {
-        // ballots: voter_id, content
         let ballots = "\
             user_A,content_A
             user_B,content_B
@@ -499,22 +495,16 @@ mod tests {
             user_D,content_D
             user_E,content_E";
 
-        // voters: voter_id, delegate_count
         let voters = "\
             user_A,1
             user_B,0
             user_D,3
             user_E,2
-            user_F,1"; // user_F has no ballot -> eligible voter
+            user_F,1";
 
         let (result, elegible_voters, ballots_without_voter, casted_ballots) =
             run_merge_join_delegates_test(ballots, voters)?;
 
-        // 10 entries in the result vector:
-        //   user_A → 1 (own) + 1 (delegate) = 2 copies
-        //   user_B → 1 (own) + 0 (delegate) = 1 copy
-        //   user_D → 1 (own) + 3 (delegate) = 4 copies
-        //   user_E → 1 (own) + 2 (delegate) = 3 copies
         assert_eq!(result.len(), 10);
         assert_eq!(
             result,
@@ -531,22 +521,13 @@ mod tests {
                 "content_E"
             ]
         );
-
-        // User_C has no matching voter: counted as “without voter”
         assert_eq!(ballots_without_voter, 1);
-
-        // Total ballots cast = 2 (A) + 1 (B) + 1 (C) + 4 (D) + 3 (E) = 11
         assert_eq!(casted_ballots, 11);
-
-        // 5 eligible voters (A, B, D, E, F)
         assert_eq!(elegible_voters, 5);
 
         Ok(())
     }
 
-    /// ------------------------------------------------------------------
-    /// 2. Empty / missing keys
-    /// ------------------------------------------------------------------
     #[test]
     fn test_missing_and_empty_keys() -> Result<()> {
         let ballots = "\
@@ -564,9 +545,6 @@ mod tests {
         let (result, elegible_voters, ballots_without_voter, casted_ballots) =
             run_merge_join_delegates_test(ballots, voters)?;
 
-        // Only user_A and user_D should be matched
-        // user_A -> 1 + 1 = 2 copies
-        // user_D -> 1 + 2 = 3 copies
         assert_eq!(result.len(), 5);
         assert_eq!(
             result,
@@ -578,22 +556,13 @@ mod tests {
                 "content_D"
             ]
         );
-
-        // Ballot 2 (empty voter id) and 3 (user_C) are "without voter"
         assert_eq!(ballots_without_voter, 2);
-
-        // 3 eligible voters (A, B, D)
         assert_eq!(elegible_voters, 3);
-
-        // Total ballots cast = 2 (A) + 1 (B) + 1 (C) + 3 (D) = 7
         assert_eq!(casted_ballots, 7);
 
         Ok(())
     }
 
-    /// ------------------------------------------------------------------
-    /// 3. Invalid delegate count
-    /// ------------------------------------------------------------------
     #[test]
     fn test_invalid_delegate_count() -> Result<()> {
         let ballots = "\
@@ -602,23 +571,15 @@ mod tests {
 
         let voters = "\
             user_A,1
-            user_G,not_a_number"; // invalid count
+            user_G,not_a_number";
 
         let (result, elegible_voters, ballots_without_voter, casted_ballots) =
             run_merge_join_delegates_test(ballots, voters)?;
 
-        // user_A → matched, 1 + 1 = 2 copies
         assert_eq!(result.len(), 2);
         assert_eq!(result, vec!["content_A", "content_A"]);
-
-        // user_G's voter record is skipped due to invalid parse.
-        // This means user_G's *ballot* is not matched.
         assert_eq!(ballots_without_voter, 1);
-
-        // Total casted ballots = 2 (A) + 1 (G) = 3
         assert_eq!(casted_ballots, 3);
-
-        // Only user_A is counted as an eligible voter. user_G was skipped.
         assert_eq!(elegible_voters, 1);
 
         Ok(())
