+++ packages/windmill/src/services/protocol_manager.rs
@@ -294,33 +294,6 @@ pub fn get_public_key_hash<C: Ctx>(messages: &Vec<Message>) -> Result<PublicKeyH
     Ok(PublicKeyHash(strand::util::to_u8_array(&pk_h)?))
 }
 
-#[instrument(skip_all)]
-pub fn generate_trustee_set<C: Ctx>(
-    configuration: &Configuration<C>,
-    trustee_pks: Vec<StrandSignaturePk>,
-) -> TrusteeSet {
-    let mut selected_trustees: TrusteeSet = [NULL_TRUSTEE; MAX_TRUSTEES];
-    let trustee_ids: Vec<usize> = trustee_pks
-        .into_iter()
-        .map(|trustee_pk| {
-            let position = configuration
-                .trustees
-                .clone()
-                .into_iter()
-                .position(|trustee| trustee == trustee_pk);
-            match position {
-                Some(value) => value + 1,
-                None => NULL_TRUSTEE,
-            }
-        })
-        .collect();
-    for i in 0..trustee_ids.len() {
-        selected_trustees[i] = trustee_ids[i];
-    }
-    event!(Level::INFO, "TrusteeSet: {:?}", selected_trustees);
-    selected_trustees
-}
-
 #[instrument(skip_all, err)]
 pub fn convert_b3(b3: &Vec<B3MessageRow>) -> Result<Vec<Message>> {
     let messages: Vec<Message> = b3
@@ -359,60 +332,6 @@ pub async fn get_b3<C: Ctx>(
     Ok(messages)
 }
 
-#[instrument(
-    skip(
-        messages,
-        configuration,
-        public_key_hash,
-        selected_trustees,
-        ballots,
-        b3_client
-    ),
-    err
-)]
-pub async fn add_ballots_to_board<C: Ctx>(
-    pm: &ProtocolManager<C>,
-    b3_client: &mut PgsqlB3Client,
-    board_name: &str,
-    messages: &Vec<Message>,
-    configuration: &Configuration<C>,
-    public_key_hash: PublicKeyHash,
-    selected_trustees: TrusteeSet,
-    ballots: Vec<Ciphertext<C>>,
-    batch: BatchNumber,
-) -> Result<()> {
-    let existing_message = messages.iter().find(|message| {
-        let batch_number = message.statement.get_batch_number();
-        let kind = message.statement.get_kind();
-        batch_number == batch && StatementType::Ballots == kind
-    });
-    if let Some(_message) = existing_message {
-        event!(
-            Level::INFO,
-            "Not adding Ballot to board {} as it already exists for batch {}",
-            board_name,
-            batch
-        );
-        return Ok(());
-    }
-
-    let ballots_len = ballots.len();
-
-    let message = Message::ballots_msg::<C, ProtocolManager<C>>(
-        configuration,
-        batch,
-        &Ballots::<C>::new(ballots),
-        selected_trustees,
-        public_key_hash,
-        pm,
-    )?;
-    info!(
-        "Adding configuration to the board for batch {} and number of ballots {}",
-        batch, ballots_len
-    );
-    b3_client.insert_ballots::<C>(board_name, message).await
-}
-
 #[instrument(err)]
 pub async fn get_board_client() -> Result<BoardClient> {
     let username = env::var("IMMUDB_USER").context("IMMUDB_USER must be set")?;
@@ -422,23 +341,6 @@ pub async fn get_board_client() -> Result<BoardClient> {
     let board_client = BoardClient::new(&server_url, &username, &password).await?;
 
     Ok(board_client)
-}
-
-#[instrument(err)]
-pub async fn get_b3_pgsql_client() -> Result<PgsqlB3Client> {
-    let username = env::var("B3_PG_USER").context("B3_PG_USER must be set")?;
-    let password = env::var("B3_PG_PASSWORD").context("B3_PG_PASSWORD must be set")?;
-    let host = env::var("B3_PG_HOST").context("B3_PG_HOST must be set")?;
-    let port = env::var("B3_PG_PORT").context("B3_PG_PORT must be set")?;
-    let database = env::var("B3_PG_DATABASE").context("B3_PG_DATABASE must be set")?;
-
-    let port: u32 = port.parse::<u32>()?;
-
-    let c = PgsqlConnectionParams::new(&host, port, &username, &password);
-    let c_db = c.with_database(&database);
-    let client = PgsqlB3Client::new(&c_db).await?;
-
-    Ok(client)
 }
 
 #[instrument(err)]
@@ -494,6 +396,23 @@ pub fn convert_board_messages(board_messages: &Vec<B3MessageRow>) -> Result<Vec<
     Ok(messages)
 }
 
+#[instrument(err)]
+pub async fn get_b3_pgsql_client() -> Result<PgsqlB3Client> {
+    let username = env::var("B3_PG_USER").context("B3_PG_USER must be set")?;
+    let password = env::var("B3_PG_PASSWORD").context("B3_PG_PASSWORD must be set")?;
+    let host = env::var("B3_PG_HOST").context("B3_PG_HOST must be set")?;
+    let port = env::var("B3_PG_PORT").context("B3_PG_PORT must be set")?;
+    let database = env::var("B3_PG_DATABASE").context("B3_PG_DATABASE must be set")?;
+
+    let port: u32 = port.parse::<u32>()?;
+
+    let c = PgsqlConnectionParams::new(&host, port, &username, &password);
+    let c_db = c.with_database(&database);
+    let client = PgsqlB3Client::new(&c_db).await?;
+
+    Ok(client)
+}
+
 pub async fn get_board_messages<C: Ctx>(
     board_name: &str,
     b3_client: &PgsqlB3Client,
@@ -502,3 +421,4 @@ pub async fn get_board_messages<C: Ctx>(
     let messages: Vec<Message> = convert_board_messages(&board_messages)?;
     Ok(messages)
 }
+
